+++ src/main/java/com/controller/QuestionController.java
@@ -17,7 +17,7 @@ public class QuestionController {
     private questionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable("id") Integer id, Model model){
+    public String question(@PathVariable("id") Long id, Model model){
         Question question = questionService.getById(id);
         // 增加浏览次数
         questionService.incView(id);
