+++ src/main/java/com/enity/Comment.java
@@ -6,13 +6,13 @@ import lombok.Data;
 
 @Data
 public class Comment {
-    private Integer id;
+    private Long id;
     private Long parentId;
     private Integer type;
     private Long commentator;
     private Long gmtCreate;
     private Long gmtModified;
     private Integer commentCount;
-    private Integer likeCount;
+    private Long likeCount;
     private String content;
 }
